backend/requirements.txt
fastapi
uvicorn
pydantic
pdfplumber
pytesseract
Pillow
sentence-transformers
scikit-learn
qdrant-client
python-dotenv